+++ src/app/[lng]/(sub)/join/funnels/step2/Step2Component.client.tsx
@@ -1,11 +1,14 @@
 'use client';
 
 import AgreeBottomSheet from './components/AgreeBottomSheet.client';
-import SchoolForm from './components/SchoolForm.client';
+import SchoolInput from './components/SchoolInput';
+import SchoolResultList from './components/SchoolResult';
 import JoinTitleTextMessage from '../../components/JoinTitleTextMessage.client';
 import { useTranslation } from '@/app/i18n/client';
+import { Loading } from '@/components/Loading';
 import { useDidMount } from '@/hooks/common/useDidMount';
 import { useModal } from '@/hooks/useModal';
+import { Suspense } from 'react';
 
 export default function Step2Component() {
   const { t } = useTranslation('join');
@@ -18,7 +21,10 @@ export default function Step2Component() {
   return (
     <main>
       <JoinTitleTextMessage>{t('chooseSchool')}</JoinTitleTextMessage>
-      <SchoolForm />
+      <SchoolInput />
+      <Suspense fallback={<Loading />}>
+        <SchoolResultList />
+      </Suspense>
     </main>
   );
 }
